recipe: Set build number to 1.
This tricks the solver into not seeing the package in `defaults`
(libtiff 4.0.6 1). Instead it will see ours. However, this package will
never be used. In fact, the solver will see this and then see the newer
version (libtiff 4.0.6 3) and upgrade outright. This requires no
additional changes to any packages that expect to use our version of
`libtiff` beyond whatever pinning is already done.